Introduce OpenStack deployment
This adds OpenStack Heat templates to launch a full edX stack in a
single node, or in a multi-node OpenStack environment with an arbitrary
number of app servers (through the use of a dynamic inventory generator).

Sample playbooks for deployment have been included, as well as sample
group variables and SSL certificates, both for the single- or multi-node
Heat templates.