Add hastexo_xblock roles
The included roles and playbook are meant install the hastexo_xblock and
its dependency, the gateone server.

Note that in addition to installing the XBlock, it must be added to the
LMS's Django INSTALLED_APPS (so it can use the edxapp Celery workers).
To do this, the included playbook depends on the edxapp role and its
variables.  To speed up the runs of edxapp (if it has already been
installed before, of course), use "--tags edxapp_cfg".

To install this in a cluster, --limit the run of the playbook to the app
servers.

Note: this depends on two modifications to the edxapp role: one that
adds support for including locations in the LMS nginx configuration, and
another that exposes the EDXAPP_ADDL_INSTALLED_APPS variable.