Fix mariadb cluster role
The mariadb cluster role expected a wrong name for the analytics API
settings variable.

Add support for the notes API database.

The old "check that this is the first server in the group" strategy also
didn't work with the current supported Ansible version.  Update it
accordinly, with "run_once".

Finally, make the MARIADB_HAPROXY_USER lines be idempotent, by using
INSERT IGNORE.