Introduce Cinder backups for MongoDB and MariaDB
This creates separate Cinder volumes for both MariaDB and MongoDB on the
backend nodes, and allows the sysadmin to take Cinder snapshots of the
mysql_data and mongodb_data volumes of a given backend instance (and
delete old ones) via a specialized playbook.

The playbook is meant to be invoked from the deployment node of a
multi-node Heat stack with the appropriate OpenStack authentication
variables properly configured.  (See backend_servers.example.)

The playbook quiesces filesystems according to each daemon's
necessities.  It does not detach volumes prior to snapshotting.

With the current Heat stack, it is intended to be called with --limit
<ip_address_of_a_secondary_backend_node>.  In other words, --limit
192.168.122.112 or --limit 192.168.122.113.  It should not be targetted
on the primary backend node, because in order to get a consistent
snapshot, prior to snapshotting it will stop the MariaDB service and
sync the filesystem.