Add backup playbook and variables
This creates a backup playbook, meant to be invoked from the deployment
node of a multi-node Heat stack with the appropriate OpenStack
authentication variables properly configured.  (See
backend_servers.example.)

As the Heat stack stands, it is intended to be called with --limit
<ip_address_of_a_secondary_backend_node>.  In other words, --limit
192.168.122.112 or --limit 192.168.122.113.  It should not be targetted
on the primary backend node, because in order to get a consistent
snapshot, prior to snapshotting it will stop the MariaDB service and
sync the filesystem.