Highly-available galera via keepalived
Introduce high-availability to galera, by using keepalived to maintain a
single VIP for the backend nodes.

On a fully-functional cluster, MySQL clients will continue to connect
exclusively to the first backend node.  In case the latter fails, they
will (re)connect to the second one.  They will only connect to backend_c
if both backend_a and backend_b fail.  This is achieved via keepalived
priority.

Keepalived will remove 3 points from a node's priority if its
MySQL service doesn't answer on TCP port 3306.

This is geared towards a 3-node scenario: in case the main node is
healthy but its MySQL service is down, the 3-point deduction guarantees
it'll drop to the last priority.